Renaming and refactoring methods in an attempt to provide sanity and make it easier to break apart Basic Edits dependency on ConfigLoader+Reader
Hopefully not too many behavioral changes, ideally none.

+++ test/test_configuration_acceptance.py
@@ -22,8 +22,8 @@ class TestConfigurationAcceptanceTest(TestCase):
             configuration_directory=self.configuration_directory)
 
     def tearDown(self):
-        shutil.rmtree(self.configuration_directory)
-
+        if os.path.isdir(self.configuration_directory):
+            shutil.rmtree(self.configuration_directory)
 
     def test_can_inject_config_directory(self):
         self.assertEquals(
